+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/blockchain/Blockchain.java
@@ -443,4 +443,11 @@ public final class Blockchain {
   public KeySetIndexProxy<HashCode> getTransactionPool() {
     return schema.getTransactionPool();
   }
+
+  /**
+   * Returns the total number of transactions committed to the blockchain.
+   */
+  public long getNumTransactions() {
+    return schema.getNumTransactions().orElse(0L);
+  }
 }